feat: add build and run scripts, environment configurations, and configuration helpers
- Added build.sh for building the Hi-Go application with environment support.
- Added run.sh for starting the application with specified environment.
- Created configuration files for dev, test, uat, and prod environments.
- Implemented configuration helpers in src/config/helpers.go for easier access to configuration values.
- Introduced legacy support in src/config/legacy.go for backward compatibility with global variables.
- Defined application configuration structure in src/config/types.go for better organization and type safety.

+++ src/service/auth_service.go
@@ -76,7 +76,7 @@ func (s *AuthService) Login(req *model.LoginRequest) (*model.LoginResponse, erro
 		User:         user,
 		AccessToken:  accessToken,
 		RefreshToken: refreshToken,
-		ExpiresIn:    config.JWTAccessTokenDuration, // 使用配置常量
+		ExpiresIn:    int64(config.JWTAccessTokenDuration), // 使用配置常量
 	}, nil
 }
